+++ bin/liftoff
@@ -3,4 +3,4 @@
 $:.push File.expand_path("../../lib", __FILE__)
 require 'liftoff'
 
-Liftoff::LaunchPad.new(ARGV).liftoff
+Liftoff::CLI.new(ARGV).run
